fix circular dependency, small changes
