+++ packages/core/src/login/webview/util.ts
@@ -3,6 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { cwQuickPickSource, cwTreeNodeSource, amazonQChatSource } from '../../codewhisperer/commands/types'
+import { ExtStartUpSources } from '../../shared/telemetry'
+import { vscodeComponent } from '../../shared/vscode/commands2'
+
 export type AuthUiClick =
     | 'auth_backButton'
     | 'auth_cancelButton'
@@ -16,3 +20,21 @@ export type AuthUiClick =
     | 'auth_codeCatalystSignIn'
     | 'auth_toolkitCloseButton'
     | 'amazonq_switchToQSignIn'
+
+/**
+ * Different places the Add Connection command could be executed from.
+ *
+ * Useful for telemetry.
+ */
+export const AuthSources = {
+    addConnectionQuickPick: 'addConnectionQuickPick',
+    firstStartup: ExtStartUpSources.firstStartUp,
+    codecatalystDeveloperTools: 'codecatalystDeveloperTools',
+    vscodeComponent: vscodeComponent,
+    cwQuickPick: cwQuickPickSource,
+    cwTreeNode: cwTreeNodeSource,
+    amazonQChat: amazonQChatSource,
+    authNode: 'authNode',
+} as const
+
+export type AuthSource = (typeof AuthSources)[keyof typeof AuthSources]
